Crossed (canonical 100): RndLine::UpdateLine(Transform,float) 90.71 ->
100 (named dist/view locals blocked the image's tail-merge of the two
clip Interp calls; new Point::ViewPos() accessor); DxShaderMgr::
LoadShaderFile 90.78 -> 100 (bases[4] is two arrays of two, unsigned k,
the k-1 test bound to a named bool); FreestyleMoveRecorder::DrawDebug
91.28 -> 100 (five file-scope statics made function-local, one anchor).

Behavioural, each verified against the target listing:
- DxShaderMgr::LoadShaderFile stored every cached pixel shader's handle
  in the vertex slot and vice versa (the k-1 polarity).
- FreestyleMoveRecorder::Poll: the depth copy wrote past the row (dst was
  not advanced per column: image stbu r11, 0x50(r5) with r5 = depthDst +
  col - 0x50 at 0x82524DBC) and kept other players' depth (the image
  zeroes unless pixelPlayer == playerIdx AND playerIdx >= 0, 0x82524DD4/
  0x82524DE0).
- HamDirector::Poll, post-proc overlay block: the force blend is tested
  against 1.0f (__real@3f800000 at 0x82478FEC), not the running blend;
  newBlend clamps to 1.0f; overlayA/overlayB are assigned in both interp
  arms; the visualizer guard reads mPlayerFreestyle (lbz 0x2bc at
  0x8247919C), not mFreestyleEnabled (0x200).
- HamListRibbon::Draw: visible-count ternary was inverted (5 vs 4);
  SetFrame in both arms.
- FlowPickOne::Activate: the last history entry is mNodes.back()
  (0x82405ED8 lwz _M_finish; lwz -0x8), not operator[](size()-1). New
  ObjPtrVec::back() in obj/Object.h, pure addition.
- RndTransformable::ApplyDynamicConstraint: kConstraintSkyBox dropped the
  summed z; the reference object is selected, not two WorldXfm() calls.
- HamDirector::ClosestMove: DataArray::Str(i) passes the owning array as
  source; non-literal practice-list entries resolved without it.

Improved: MoveGraph::FindVariantPair 92.56 -> 98.10, UIListState::Scroll
92.52 -> 96.15, RndTransformable::ApplyDynamicConstraint 92.85 -> 95.88,
HamListRibbon::Draw 92.13 -> 94.96, FreestyleMoveRecorder::Poll 92.53 ->
94.55, RndLight::Projection 91.63 -> 93.32, FlowPickOne::Activate 92.93
-> 94.31, HamDirector::Poll 93.71 -> 94.89, HamDirector::ClosestMove
92.25 -> 92.66, RndBitmap::LoadDIB 92.97 -> 93.21 (biCompression is long).

Unchanged with the residual pinpointed: GranularSynth ctor 91.9 (we hold
pi in a fifth callee-saved FPR; the image reloads it from .rdata inside
the loop, two spellings exactly inert), MemHeap::Print 91.8 (two
negatives, one regressed to 89.8), HamRibbon::ConstructMesh 90.2,
SpectralAnalysis::SetMode 91.7, LoopVizCallback::UpdateOverlay (operand
order applied; the fadds f28,f29 label position recorded, not applied
because it perturbs FPR allocation), DxMesh::OnSync (two refuted
spellings, MakeString row is an ICF fold).

Lane finding, four independent confirmations: a pure declaration reorder
inside one statement or block is inert; what moves codegen is a
declaration crossing a call, removing a local so the value is re-read,
or statement order across two assignments in one basic block.

Lane ran as Opus. Headers rndobj/Bitmap.h, rndobj/Line.h, obj/Object.h
each measured with a full ninja; final compare_progress vs base
130e2efbb: 13 improvements, 0 regressions.

+++ src/system/rndobj/Bitmap.cpp
@@ -1183,8 +1183,13 @@ bool RndBitmap::LoadDIB(BinStream *bs, unsigned int offbits) {
         return false;
     }
     if (infoheader.biCompression != 0) {
+        // The target binds MakeString's `const long &` straight to the field's
+        // own stack slot: 0x82672CAC `addi r5, r1, 0x80`, and infoheader lives
+        // at 0x70(r1). A `(long)` value cast spills a second copy at 0x60(r1)
+        // and pushes bs->Name() out to 0x64; a `(const long &)` cast does the
+        // same. Only a `long` member binds the reference to the field itself.
         MILO_NOTIFY(
-            "%s: Unsupported compression %d", bs->Name(), (long)infoheader.biCompression
+            "%s: Unsupported compression %d", bs->Name(), infoheader.biCompression
         );
         return false;
     }
